Moving test files to test folder.
