+++ test/llm.test.js
@@ -1,6 +1,5 @@
-import { jest } from '@jest/globals';
-import game from '../../__mocks__/game.js';
-import { getGenerationParameters, llmParametersAndDefaults } from './llm.js';
+import game from '../__mocks__/game.js';
+import { getGenerationParameters, llmParametersAndDefaults } from '../src/scripts/llm.js';
 
 describe('getGenerationParameters', () => {
     beforeEach(() => {
